feat(34-02): split features page into server wrapper and client
